+++ src/sll/stack.c
@@ -17,14 +17,14 @@ sll_assembly_instruction_t* _acquire_next_instruction(sll_assembly_data_t* a_dt)
 		*((void**)(a_dt->_s.e))=n;
 		*((void**)n)=NULL;
 		sll_assembly_instruction_t* s=(sll_assembly_instruction_t*)((char*)n+sizeof(void*));
-		s->t=ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
+		s->t=SLL_ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
 		s->dt._p=a_dt->_s.p-1;
-		SLL_ASSERT(a_dt->_s.p->t==ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK);
+		SLL_ASSERT(a_dt->_s.p->t==SLL_ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK);
 		a_dt->_s.p->dt._p=s+1;
 		a_dt->_s.c=((SLL_ROUND_PAGE(ASSEMBLY_INSTRUCTION_STACK_ALLOC_SIZE)-sizeof(void*)-sizeof(sll_assembly_instruction_t)*2)/sizeof(sll_assembly_instruction_t));
 		a_dt->_s.p=s+1;
 		s+=a_dt->_s.c+1;
-		s->t=ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
+		s->t=SLL_ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
 		s->dt._p=NULL;
 		a_dt->_s.e=n;
 	}
@@ -43,14 +43,14 @@ sll_node_t* _acquire_next_node(sll_compilation_data_t* c_dt){
 		*((void**)(c_dt->_s.e))=n;
 		*((void**)n)=NULL;
 		sll_node_t* s=(sll_node_t*)((char*)n+sizeof(void*));
-		s->t=NODE_TYPE_CHANGE_STACK;
+		s->t=SLL_NODE_TYPE_CHANGE_STACK;
 		s->dt._p=c_dt->_s.p-1;
-		SLL_ASSERT(c_dt->_s.p->t==NODE_TYPE_CHANGE_STACK);
+		SLL_ASSERT(c_dt->_s.p->t==SLL_NODE_TYPE_CHANGE_STACK);
 		c_dt->_s.p->dt._p=s+1;
 		c_dt->_s.c=((SLL_ROUND_PAGE(NODE_STACK_ALLOC_SIZE)-sizeof(void*)-sizeof(sll_node_t)*2)/sizeof(sll_node_t));
 		c_dt->_s.p=s+1;
 		s+=c_dt->_s.c+1;
-		s->t=NODE_TYPE_CHANGE_STACK;
+		s->t=SLL_NODE_TYPE_CHANGE_STACK;
 		s->dt._p=NULL;
 		c_dt->_s.e=n;
 	}
@@ -88,12 +88,12 @@ void _init_assembly_stack(sll_assembly_data_t* a_dt){
 	a_dt->_s.e=a_dt->_s.s;
 	*((void**)(a_dt->_s.s))=NULL;
 	sll_assembly_instruction_t* s=(sll_assembly_instruction_t*)((char*)(a_dt->_s.s)+sizeof(void*));
-	s->t=ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
+	s->t=SLL_ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
 	s->dt._p=NULL;
 	a_dt->_s.c=((SLL_ROUND_PAGE(ASSEMBLY_INSTRUCTION_STACK_ALLOC_SIZE)-sizeof(void*)-sizeof(sll_assembly_instruction_t)*2)/sizeof(sll_assembly_instruction_t));
 	a_dt->_s.p=s+1;
 	s+=a_dt->_s.c+1;
-	s->t=ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
+	s->t=SLL_ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
 	s->dt._p=NULL;
 }
 
@@ -104,12 +104,12 @@ void _init_node_stack(sll_compilation_data_t* c_dt){
 	c_dt->_s.e=c_dt->_s.s;
 	*((void**)(c_dt->_s.s))=NULL;
 	sll_node_t* s=(sll_node_t*)((char*)(c_dt->_s.s)+sizeof(void*));
-	s->t=NODE_TYPE_CHANGE_STACK;
+	s->t=SLL_NODE_TYPE_CHANGE_STACK;
 	s->dt._p=NULL;
 	c_dt->_s.c=((SLL_ROUND_PAGE(NODE_STACK_ALLOC_SIZE)-sizeof(void*)-sizeof(sll_node_t)*2)/sizeof(sll_node_t));
 	c_dt->_s.p=s+1;
 	s+=c_dt->_s.c+1;
-	s->t=NODE_TYPE_CHANGE_STACK;
+	s->t=SLL_NODE_TYPE_CHANGE_STACK;
 	s->dt._p=NULL;
 }
 
@@ -120,10 +120,10 @@ void _shift_nodes(sll_node_t* o,sll_compilation_data_t* c_dt,sll_node_offset_t o
 		return;
 	}
 	sll_node_t* s=c_dt->_s.p-1;
-	if (o->t==NODE_TYPE_CHANGE_STACK){
+	if (o->t==SLL_NODE_TYPE_CHANGE_STACK){
 		o=o->dt._p;
 	}
-	if (s->t==NODE_TYPE_CHANGE_STACK){
+	if (s->t==SLL_NODE_TYPE_CHANGE_STACK){
 		s=s->dt._p;
 	}
 	for (sll_node_offset_t i=1;i<off;i++){
@@ -136,11 +136,11 @@ void _shift_nodes(sll_node_t* o,sll_compilation_data_t* c_dt,sll_node_offset_t o
 			(*(c_dt->ft.dt+d->dt.fn.id))->off+=off;
 		}
 		s--;
-		if (s->t==NODE_TYPE_CHANGE_STACK){
+		if (s->t==SLL_NODE_TYPE_CHANGE_STACK){
 			s=s->dt._p;
 		}
 		d--;
-		if (d->t==NODE_TYPE_CHANGE_STACK){
+		if (d->t==SLL_NODE_TYPE_CHANGE_STACK){
 			d=d->dt._p;
 		}
 	} while (s!=o);